A single script can now serve many buttons
All button, input and a tags with a class of `subtome` will be clickable if the script is loaded with a `standalone` parameter like:

<script type="text/javascript">
  (function() {
    var po = document.createElement('script'); po.type = 'text/javascript'; po.async = true;
    po.src = 'http://127.0.0.1:3000/load.js#standalone';
    var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(po, s);
  })();
</script>

Button as well as the script can have a `resource` and a `feed` parameter. The button defines them through data attributes like `data-resource` and the script using query parameters in the hash, like `http://example.com/load.js#resource=/about&feed=/feed.xml`. The parameters of the individual button overrides the global one of the script.

The script is still backwards compatible with the old one - for those button it works like usual.